% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CDpca.R
\name{CDpca}
\alias{CDpca}
\title{CDpca}
\usage{
CDpca(data, class, fixAtt, nnloads = 0, Q, P, tol, maxit, r)
}
\arguments{
\item{data}{data frame (numeric).}

\item{class}{vector (numeric) or 0, if classes of objects are unknown.}

\item{fixAtt}{vector (numeric) or 0, for a selection of attributes.}

\item{nnloads}{1 or 0, for nonnegative loadings.}

\item{Q}{integer, number of clusters of variables.}

\item{P}{integer, number of clusters of objects.}

\item{tol}{real number, small positive tolerance.}

\item{maxit}{integer, maximum of iterations.}

\item{r}{number of runs of the cdpca algoritm for the final solution.}
}
\value{
iter: iterations used in the best loop for computing the best solution
 loop: best loop number
 timebestloop: computation time on the best loop
 timeallloops: computation time for all loops
 Y: the component score matrix
 Ybar: the object centroids matrix in the reduced space
 A: the component loading matrix
 U: the partition of objects
 V: the partition of variables
 Fobj: function to maximize
 bcdev: between cluster deviance
 bcdevTotal: between cluster deviance over the total variability
 tableclass: cdpca classification
 pseudocm: pseudo confusion matrix of the real and cdpca classifications
 Enorm: error norm for the obtained cdpca model
}
\description{
CDpca performs a clustering and disjoint principal components analysis
on the given numeric data matrix and returns a list of results.
}
\examples{
exampleCDpca = CDpca(data, class, fixAtt, nnloads, Q, P, tol, maxit, r)
}
\keyword{cdpca}
\keyword{cluster}
\keyword{kmeans}
\keyword{pca}
