% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CDpcaplot.R
\name{CDpcaplot}
\alias{CDpcaplot}
\title{CDpcaplot}
\usage{
CDpcaplot(data, obj, class = 0, save = 0, filename)
}
\arguments{
\item{data}{data frame (numeric).}

\item{obj}{CDpca object.}

\item{class}{Real classification for comparison (optional).}
}
\description{
CDpcaplot plots CDpca object.
}
\examples{
exampleplot = CDpcaplot(data, obj, class)
}
\keyword{cdpca}
\keyword{cluster}
\keyword{kmeans}
\keyword{pca}
