% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CDpcaSummary.R
\name{CDpcaSummary}
\alias{CDpcaSummary}
\title{CDpcaSummary}
\usage{
CDpcaSummary(obj)
}
\arguments{
\item{obj}{CDpca object}
}
\description{
Print a variety of useful information regarding a CDpca object.
}
