% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NestedCDpca.R
\name{NestedCDpca}
\alias{NestedCDpca}
\title{NestedCDpca}
\usage{
NestedCDpca(data, method = "CDpca", class, fixAtt, nnloads = 0, Q, P,
  K, ent = 1, tol, maxit, r, stand)
}
\arguments{
\item{data}{data frame (numeric).}

\item{method}{method to apply. CDpca or RCDpca.}

\item{class}{vector (numeric) or 0, if classes of objects are unknown.}

\item{fixAtt}{vector (numeric) or 0, for a selection of attributes.}

\item{nnloads}{1 or 0, for nonnegative loadings}

\item{Q}{integer, number of clusters of variables.}

\item{P}{integer, number of clusters of objects.}

\item{K}{vector (numeric), representing number of sub-clusters.}

\item{ent}{integer, fuzzifier parameter (default 1).}

\item{tol}{real number, small positive tolerance.}

\item{maxit}{integer, maximum of iterations.}

\item{r}{number of runs of the cdpca algoritm for the final solution.}

\item{stand}{integer, 1 to standardize data before applying relaxed FKM (0 otherwise).}
}
\value{
firstCDpca: CDpca object.
 secondCDpca: CDpca objects of second run of CDpca.
}
\description{
NestedCDpca performs (sequentially) two clustering and disjoint principal components analysis
on the given numeric data matrix and returns a list of results.
}
\examples{
example = NestedCDpca(data, method, class, fixAtt, nnloads, Q, P, K, ent, tol, maxit, r)
}
\keyword{cdpca}
\keyword{cluster}
\keyword{kmeans}
\keyword{pca}
