% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NestedCDpcaplot.R
\name{NestedCDpcaplot}
\alias{NestedCDpcaplot}
\title{NestedCDpcaplot}
\usage{
NestedCDpcaplot(data, obj, class = 0, save = 0, filename)
}
\arguments{
\item{data}{data frame (numeric).}

\item{obj}{NestedCDpca object.}

\item{class}{Real classification for comparison (optional).}
}
\description{
NestedCDpcaplot plots NestedCDpca object.
}
\examples{
exampleplot = NestedCDpcaplot(data, obj, class)
}
\keyword{cdpca}
\keyword{cluster}
\keyword{kmeans}
\keyword{pca}
