% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NestedCDpcaSummary.R
\name{NestedCDpcaSummary}
\alias{NestedCDpcaSummary}
\title{NestedCDpcaSummary}
\usage{
NestedCDpcaSummary(obj)
}
\arguments{
\item{obj}{NestedCDpca object}
}
\description{
Print a variety of useful information regarding a NestedCDpca object.
}
