% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PMlargestEigen.R
\name{PMlargestEigen}
\alias{PMlargestEigen}
\title{PMlargestEigen}
\usage{
PMlargestEigen(M)
}
\arguments{
\item{M}{covariance matrix restricted to some attributes (symmetric)}
}
\description{
The Power Method simple function
}
