% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RelaxedCDpca.R
\name{RelaxedCDpca}
\alias{RelaxedCDpca}
\title{RelaxedCDpca}
\usage{
RelaxedCDpca(data, class, fixAtt, nnloads = 0, Q, P, ent = 1, tol,
  maxit, r, stand = 0)
}
\arguments{
\item{data}{data frame (numeric).}

\item{class}{vector (numeric) or 0, if classes of objects are unknown.}

\item{fixAtt}{vector (numeric) or 0, for a selection of attributes.}

\item{nnloads}{1 or 0, for nonnegative loadings}

\item{Q}{integer, number of clusters of variables.}

\item{P}{integer, number of clusters of objects.}

\item{ent}{integer, fuzzifier parameter (default 1).}

\item{tol}{real number, small positive tolerance.}

\item{maxit}{integer, maximum of iterations.}

\item{r}{number of runs of the cdpca algoritm for the final solution.}

\item{stand}{integer, 1 to standardize data before applying relaxed FKM (0 otherwise).}
}
\value{
iter: iterations used in the best loop for computing the best solution
 loop: best loop number
 timebestloop: computation time on the best loop
 timeallloops: computation time for all loops
 Y: the component score matrix
 Ybar: the object centroids matrix in the reduced space
 A: the component loading matrix
 U: the partition of objects
 V: the partition of variables
 Fobj: function to maximize
 bcdev: between cluster deviance
 bcdevTotal: between cluster deviance over the total variability
 tableclass: cdpca classification
 pseudocm: pseudo confusion matrix of the real and cdpca classifications
 Enorm: error norm for the obtained cdpca model
}
\description{
RelaxedCDpca performs a relaxed clustering and disjoint principal components analysis
on the given numeric data matrix and returns a list of results.
}
\examples{
exampleRelaxed = RelaxedCDpca(data, class, fixAtt, nnloads, Q, P, ent, tol, maxit, r)
}
\keyword{cdpca}
\keyword{cluster}
\keyword{kmeans}
\keyword{pca}
