% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RelaxedCDpcaplot.R
\name{RelaxedCDpcaplot}
\alias{RelaxedCDpcaplot}
\title{RelaxedCDpcaplot}
\usage{
RelaxedCDpcaplot(data, obj, class = 0, save = 0, filename)
}
\arguments{
\item{data}{data frame (numeric).}

\item{obj}{RelaxedCDpca object.}

\item{class}{Real classification for comparison (optional).}
}
\description{
RelaxedCDpcaplot plots RelaxedCDpca object.
}
\examples{
exampleplot = RelaxedCDpcaplot(data, obj, class)
}
\keyword{cdpca}
\keyword{cluster}
\keyword{kmeans}
\keyword{pca}
