% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RelaxedCDpcaSummary.R
\name{RelaxedCDpcaSummary}
\alias{RelaxedCDpcaSummary}
\title{RelaxedCDpcaSummary}
\usage{
RelaxedCDpcaSummary(obj)
}
\arguments{
\item{obj}{RelaxedCDpca object}
}
\description{
Print a variety of useful information regarding a RelaxedCDpca object.
}
