% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SDPinitialization.R
\name{SDPinitialization}
\alias{SDPinitialization}
\title{SDPinitialization}
\usage{
SDPinitialization(dim1, dim2, D)
}
\arguments{
\item{dim1}{number of rows}

\item{dim2}{number of columns}
}
\value{
The random matrix (dim1xdim2) with only one nonzero element per row.
}
\description{
Generates a binary and row stochastic matrix using SDP-based approach
}
\examples{
SDPinitialization(1, 2, data)
}
